requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
google-api-python-client
google-auth-oauthlib
python-dateutil
